ASPlayer: CF2 add video/audio render first frame callback [1/1]
PD#SWPL-123148

Problem:
add video/audio render first frame callback

Solution:
add video/audio render first frame callback

Verify:
AH212 Android T

+++ libs/ASPlayer-library/src/main/java/com/amlogic/asplayer/core/InputBuffer.java
@@ -31,7 +31,7 @@ public class InputBuffer {
     public int error;
 
     // additional details for video
-    public float pixelAspectRatio;
+    public int pixelAspectRatio;
     public int width;
     public int height;
 }
